PR comments: Move shim logic into its own module
