+++ Public/Src/Sandbox/Windows/DetoursServices/DetouredFunctions.cpp
@@ -16,6 +16,7 @@
 #include "UnicodeConverter.h"
 #include "MetadataOverrides.h"
 #include "HandleOverlay.h"
+#include "SubstituteProcessExecution.h"
 
 using std::wstring;
 using std::unique_ptr;
@@ -1778,230 +1779,6 @@ NTSTATUS NTAPI Detoured_ZwSetInformationFile(
         FileInformationClass);
 }
 
-/// Runs an injected AnyBuild shim instead of a child process, passing the
-/// original command and arguments to the shim along with, implicitly,
-/// the current working directory and environment.
-BOOL WINAPI InjectShim(
-    _In_opt_    LPCWSTR               lpApplicationName,
-    _Inout_opt_ LPWSTR                lpCommandLine,
-    _In_opt_    LPSECURITY_ATTRIBUTES lpProcessAttributes,
-    _In_opt_    LPSECURITY_ATTRIBUTES lpThreadAttributes,
-    _In_        BOOL                  bInheritHandles,
-    _In_        DWORD                 dwCreationFlags,
-    _In_opt_    LPVOID                lpEnvironment,
-    _In_opt_    LPCWSTR               lpCurrentDirectory,
-    _In_        LPSTARTUPINFOW        lpStartupInfo,
-    _Out_       LPPROCESS_INFORMATION lpProcessInformation)
-{
-    // Create a final buffer for the original command line.
-    size_t fullCmdLineSizeInChars =
-        (lpApplicationName != nullptr ? wcslen(lpApplicationName) : 0) +
-        (lpCommandLine != nullptr ? wcslen(lpCommandLine) : 0) +
-        4;  // Quotes and space and trailing null
-    wchar_t *fullCommandLine = new wchar_t[fullCmdLineSizeInChars];
-    if (fullCommandLine == nullptr)
-    {
-        Dbg(L"Failure running AnyBuild shim process - failed to allocate buffer of size %d.", fullCmdLineSizeInChars * sizeof(WCHAR));
-        SetLastError(ERROR_OUTOFMEMORY);
-        return FALSE;
-    }
-
-    fullCommandLine[0] = L'\0';
-    if (lpApplicationName != nullptr)
-    {
-        fullCommandLine[0] = L'"';
-        wcscpy_s(fullCommandLine + 1, fullCmdLineSizeInChars, lpApplicationName);
-        wcscat_s(fullCommandLine, fullCmdLineSizeInChars, L"\" ");
-    }
-    if (lpCommandLine != nullptr)
-    {
-        wcscat_s(fullCommandLine, fullCmdLineSizeInChars, lpCommandLine);
-    }
-
-    Dbg(L"Injecting AnyBuild shim '%s' for process command line '%s'", g_substituteProcessExecutionShimPath, fullCommandLine);
-    BOOL rv = Real_CreateProcessW(
-        /*lpApplicationName:*/ g_substituteProcessExecutionShimPath,
-        /*lpCommandLine:*/ fullCommandLine,
-        lpProcessAttributes,
-        lpThreadAttributes,
-        bInheritHandles,
-        dwCreationFlags,
-        lpEnvironment,
-        lpCurrentDirectory,
-        lpStartupInfo,
-        lpProcessInformation);
-
-    delete[] fullCommandLine;
-    return rv;
-}
-
-// https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
-inline const wchar_t *trim_start(const wchar_t *str)
-{
-    while (wmemchr(L" \t\n\r", *str, 4))  ++str;
-    return str;
-}
-
-// https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
-inline const wchar_t *trim_end(const wchar_t *end)
-{
-    while (wmemchr(L" \t\n\r", end[-1], 4)) --end;
-    return end;
-}
-
-// https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
-inline std::wstring trim(const wchar_t *buffer, size_t len) // trim a buffer (input?)
-{
-    return std::wstring(trim_start(buffer), trim_end(buffer + len));
-}
-
-// https://stackoverflow.com/questions/216823/whats-the-best-way-to-trim-stdstring
-inline void trim_inplace(std::wstring& str)
-{
-    str.assign(trim_start(str.c_str()),
-        trim_end(str.c_str() + str.length()));
-}
-
-// Returns new value for lpCommandLine pointing to the remainder of the string
-// to use as the command line parameters.
-static const void FindApplicationNameFromCommandLine(const wchar_t *lpCommandLine, _Out_ wstring &command, _Out_ wstring &commandArgs)
-{
-    wstring fullCommandLine(lpCommandLine);
-    if (fullCommandLine.length() == 0)
-    {
-        command = wstring();
-        commandArgs = wstring();
-        return;
-    }
-
-    if (fullCommandLine[0] == L'"')
-    {
-        // Find the close quote. Might not be present which means the command
-        // is the full command line minus the initial quote.
-        size_t closeQuoteIndex = fullCommandLine.find('"', 1);
-        if (closeQuoteIndex >= 1)
-        {
-            if (closeQuoteIndex == fullCommandLine.length() - 1)
-            {
-                // Quotes cover entire command line.
-                command = fullCommandLine.substr(1, fullCommandLine.length() - 2);
-                trim_inplace(command);
-                commandArgs = wstring();
-            }
-            else
-            {
-                wstring noQuoteCommand = fullCommandLine.substr(1, closeQuoteIndex - 1);
-
-                // Find the next delimiting space after the close double-quote.
-                // For example a command line like "c:\program files"\foo we need to
-                // keep \foo and cut the quotes to produce c:\program files\foo
-                size_t spaceDelimiterIndex = fullCommandLine.find(L' ', closeQuoteIndex + 1);
-                if (spaceDelimiterIndex == wstring::npos)
-                {
-                    // No space, take everything through the end of the command line.
-                    spaceDelimiterIndex = fullCommandLine.length();
-                }
-
-                command = (noQuoteCommand +
-                    fullCommandLine.substr(closeQuoteIndex + 1, spaceDelimiterIndex - closeQuoteIndex));
-                trim_inplace(command);
-                commandArgs = fullCommandLine.substr(spaceDelimiterIndex + 1);
-                trim_inplace(commandArgs);
-            }
-        }
-        else
-        {
-            // No close quote. Take everything through the end of the command line as the command.
-            command = fullCommandLine.substr(1);
-            commandArgs = wstring();
-        }
-    }
-    else
-    {
-        // No open quote, pure space delimiter.
-        size_t spaceDelimiterIndex = fullCommandLine.find(' ');
-        if (spaceDelimiterIndex == wstring::npos)
-        {
-            // No space, take everything through the end of the command line.
-            spaceDelimiterIndex = fullCommandLine.length();
-        }
-
-        command = fullCommandLine.substr(0, spaceDelimiterIndex);
-        commandArgs = fullCommandLine.substr(spaceDelimiterIndex + 1);
-        trim_inplace(commandArgs);
-    }
-}
-
-static bool CommandArgsContainMatch(const wchar_t *commandArgs, const wchar_t *argMatch)
-{
-    if (argMatch == nullptr)
-    {
-        // No optional match, meaning always match.
-        return true;
-    }
-
-    return wcsstr(commandArgs, argMatch) != nullptr;
-}
-
-static bool ShouldShim(const wstring &command, const wchar_t *commandArgs)
-{
-    // Easy cases.
-    if (g_pShimProcessMatches == nullptr || g_pShimProcessMatches->empty())
-    {
-        // Shim everything or shim nothing if there are no matches to compare.
-        return g_ProcessExecutionShimAllProcesses;
-    }
-
-    size_t commandLen = command.length();
-
-    bool foundMatch = false;
-
-    for (auto it = g_pShimProcessMatches->begin(); it != g_pShimProcessMatches->end(); ++it)
-    {
-        ShimProcessMatch *pMatch = *it;
-
-        const wchar_t *processName = pMatch->ProcessName.get();
-        size_t processLen = wcslen(processName);
-
-        // lpAppName is longer than e.g. "cmd.exe", see if lpAppName ends with e.g. "\cmd.exe"
-        if (processLen < commandLen)
-        {
-            if (command[commandLen - processLen - 1] == L'\\' &&
-                _wcsicmp(command.c_str() + commandLen - processLen, processName) == 0)
-            {
-                if (CommandArgsContainMatch(commandArgs, pMatch->ArgumentMatch.get()))
-                {
-                    foundMatch = true;
-                    break;
-                }
-            }
-
-            continue;
-        }
-
-        if (processLen == commandLen)
-        {
-            if (_wcsicmp(processName, command.c_str()) == 0)
-            {
-                if (CommandArgsContainMatch(commandArgs, pMatch->ArgumentMatch.get()))
-                {
-                    foundMatch = true;
-                    break;
-                }
-            }
-        }
-    }
-
-    if (g_ProcessExecutionShimAllProcesses)
-    {
-        // A match means we don't want to shim - an opt-out list.
-        return !foundMatch;
-    }
-
-    // An opt-in list, shim if matching.
-    return foundMatch;
-}
-
 IMPLEMENTED(Detoured_CreateProcessW)
 BOOL WINAPI Detoured_CreateProcessW(
     _In_opt_    LPCWSTR               lpApplicationName,
@@ -2015,38 +1792,22 @@ BOOL WINAPI Detoured_CreateProcessW(
     _In_        LPSTARTUPINFOW        lpStartupInfo,
     _Out_       LPPROCESS_INFORMATION lpProcessInformation)
 {
-    if (g_substituteProcessExecutionShimPath != nullptr)
+    bool injectedShim = false;
+    BOOL ret = MaybeInjectSubstituteProcessShim(
+        lpApplicationName,
+        lpCommandLine,
+        lpProcessAttributes,
+        lpThreadAttributes,
+        bInheritHandles,
+        dwCreationFlags,
+        lpEnvironment,
+        lpCurrentDirectory,
+        lpStartupInfo,
+        lpProcessInformation,
+        injectedShim);
+    if (injectedShim)
     {
-        wstring command;
-        wstring commandArgs;
-        const wchar_t *lpCommandArgs;
-        if (lpApplicationName == nullptr)
-        {
-            FindApplicationNameFromCommandLine(lpCommandLine, command, commandArgs);
-            lpCommandArgs = commandArgs.c_str();
-        }
-        else
-        {
-            command = wstring(lpApplicationName);
-            lpCommandArgs = lpCommandLine;
-        }
-
-        if (ShouldShim(command, lpCommandArgs))
-        {
-            // Instead of Detouring the child, run the requested shim
-            // passing the original command line, but only for appropriate commands.
-            return InjectShim(
-                lpApplicationName,
-                lpCommandLine,
-                lpProcessAttributes,
-                lpThreadAttributes,
-                bInheritHandles,
-                dwCreationFlags,
-                lpEnvironment,
-                lpCurrentDirectory,
-                lpStartupInfo,
-                lpProcessInformation);
-        }
+        return ret;
     }
 
     if (!MonitorChildProcesses())
