+++ Public/Src/Sandbox/Windows/DetoursServices/DetoursServices.dsc
@@ -34,7 +34,8 @@ namespace Core {
         f`PolicySearch.h`,
         f`DeviceMap.h`,
         f`DetouredProcessInjector.h`,
-        f`UniqueHandle.h`
+        f`UniqueHandle.h`,
+        f`SubstituteProcessExecution.h`,
     ];
 
     export const pathToDeviceMapLib: PathAtom = a`${qualifier.platform.replace("x", qualifier.configuration)}`;
@@ -82,6 +83,7 @@ namespace Core {
                 f`DeviceMap.cpp`,
                 f`SendReport.cpp`,
                 f`DetouredProcessInjector.cpp`,
+                f`SubstituteProcessExecution.cpp`,
             ],
 
             exports: [
@@ -123,6 +125,7 @@ namespace Core {
                 f`PolicySearch.cpp`,
                 f`DeviceMap.cpp`,
                 f`DetouredProcessInjector.cpp`,
+                f`SubstituteProcessExecution.cpp`,
             ],
 
             exports: [
